refactor: replace float for double in expense struct

+++ include/expense.h
@@ -5,11 +5,11 @@
 struct expense {
   date date;
   std::string category;
-  float amount;
+  double amount;
   std::string desc;
 
   explicit expense();
-  explicit expense(struct date date, std::string category, float amount,
+  explicit expense(struct date date, std::string category, double amount,
                    std::string desc);
 
   std::string to_string() const;
