+++ src/expense.cpp
@@ -3,9 +3,9 @@
 #include <string>
 #include <utility>
 
-expense::expense() : date(0, 0, 0), amount(0.0F) {}
+expense::expense() : date(0, 0, 0), amount(0.0) {}
 
-expense::expense(struct date date, std::string category, float amount,
+expense::expense(struct date date, std::string category, double amount,
                  std::string desc)
     : date(date), category(std::move(category)), amount(amount),
       desc(std::move(desc)) {}
